textatlas_cn/requirements.txt
pillow>=10.0
opencv-python>=4.9
numpy>=1.26
pymupdf>=1.24
python-pptx>=0.6.23
pyyaml>=6.0
tqdm>=4.66
requests>=2.31
httpx>=0.27
pydantic>=2.6
pandas>=2.2
diskcache>=5.6
tenacity>=8.2
jieba>=0.42
fasttext-langdetect>=1.0.5
sentence-transformers>=2.7
simhash>=2.1
openai>=1.30
dashscope>=1.20
zhipuai>=2.0
huggingface_hub>=0.23
datasets>=2.20
webdataset>=0.2.86
paddlepaddle>=2.6
paddleocr>=2.7
easyocr>=1.7
cn_ocr>=2.3
ultralytics>=8.2
segment-anything>=1.0
